requirements.txt
web3
pyyaml
colorama